update true or false question answer content
